+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-07-11 22:04:44
-// Commit: feat(ui): add async ui stream provider
+// Updated at: 2026-07-12 10:01:34
+// Commit: chore(ui): update dependencies and cleanup unused imports in ui
 
 class uiComponentState {
-  final String commitRef = '2717';
-  final String lastMessage = 'feat(ui): add async ui stream provider';
-  final DateTime timestamp = DateTime.parse('2026-07-11T22:04:44+05:00');
+  final String commitRef = '2726';
+  final String lastMessage = 'chore(ui): update dependencies and cleanup unused imports in ui';
+  final DateTime timestamp = DateTime.parse('2026-07-12T10:01:34+05:00');
 }